Move bulk notes logic to `bulk_notes_util`
